Move expected and scope_guard to util subdir
