+++ src/aziomq/util/scope_guard.hpp
