+++ src/aziomq/version.hpp
@@ -1,0 +1,19 @@
+/*
+    Copyright (c) 2013-2014 Contributors as noted in the AUTHORS file
+
+    This file is part of aziomq
+
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+    This is an implementation of Andrei Alexandrescu's Expected<T> type
+    from the "Systematic Error Handling in C++" talk given at C++ And
+    Beyond 2012
+*/
+#ifndef AZIOMQ_VERSION_HPP_
+#define AZIOMQ_VERSION_HPP_
+
+// corresponds to the ZeroMQ version supported
+#define AZIOMQ_VERSION_MAJOR 4
+#define AZIOMQ_VERSION_MINOR 0
+#endif // AZIOMQ_VERSION_HPP_
