+++ src/lib/io_service.cpp
@@ -6,6 +6,7 @@
     Distributed under the Boost Software License, Version 1.0. (See accompanying
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 */
+#include "../aziomq/version.hpp"
 #include "../aziomq/detail/tracked_op.hpp"
 #include "../aziomq/detail/zeromq_socket_service.hpp"
 #include "../aziomq/io_service.hpp"
